requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
grpcio==1.46.3
grpcio-tools==1.46.3
idna==3.7
protobuf==3.20.3
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
requests==2.32.3
six==1.16.0
sounddevice==0.4.7
srt==3.5.3
tqdm==4.66.4
urllib3==2.2.1
vosk==0.3.45
Wave==0.0.2
websockets==12.0
yandex-speechkit==1.3.3
